UPDT: Cambio tipo de datos categorias

+++ src/models/Categoria.Model.js
@@ -28,9 +28,9 @@ Categoria.init({
         defaultValue: 'NINGUNO'
     },
     impuesto: {
-        type: DataTypes.DECIMAL(10, 2),
+        type: DataTypes.INTEGER,
         allowNull: true,
-        defaultValue: 0.00
+        defaultValue: 0
     }
 },
 {
